added ResponseEntity and Validation

+++ src/main/java/com/example/webshop/restUi/RestWebShopController.java
@@ -3,7 +3,10 @@ package com.example.webshop.restUi;
 import com.example.webshop.business.Category;
 import com.example.webshop.business.Product;
 import com.example.webshop.business.WebShopService;
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -20,8 +23,13 @@ public class RestWebShopController {
     }
 
     @GetMapping("/rest/bycategory")
-    public List<Product> getProductByCategory(@RequestParam Category category) {
-        return webShopService.getProductByCategory(category);
+    public ResponseEntity<List<Product>> getProductByCategory(@RequestParam Category category) {
+        if (webShopService.getProductByCategory(category).isEmpty()) {
+            return ResponseEntity.badRequest().body(null);
+        } else {
+            webShopService.getProductByCategory(category);
+            return ResponseEntity.accepted().body(webShopService.getProductByCategory(category));
+        }
     }
 
     @GetMapping("/rest/byname/{name}")
@@ -30,24 +38,45 @@ public class RestWebShopController {
     }
 
     @GetMapping("/rest/byid/{id}")
-    public Product getProductById(@PathVariable Long id) {
-        return webShopService.getProductById(id);
+    public ResponseEntity<String> getProductById(@PathVariable Long id) {
+        if (webShopService.getProductRepo().findById(id).isEmpty()) {
+            return ResponseEntity.notFound().build();
+        } else {
+            webShopService.getProductById(id);
+            return ResponseEntity.accepted().body("Här är din " + webShopService.getProductById(id));
+        }
+
     }
 
     @DeleteMapping("/rest/deleteproduct/{id}")
-    public List<Product> deleteProductById(@PathVariable Long id) {
-        return webShopService.deleteProductById(id);
+    public ResponseEntity<String> deleteProductById(@PathVariable Long id) {
+        if (webShopService.getProductRepo().findById(id).isEmpty()) {
+            return ResponseEntity.notFound().build();
+        } else {
+            webShopService.deleteProductById(id);
+            return ResponseEntity.ok("Produkten är borttagen!");
+        }
+
     }
 
     @PostMapping("/rest/addproduct")
-    public List<Product> addProduct(@RequestParam String name, @RequestParam double price, @RequestParam Category category) {
-        webShopService.add(new Product(category, name, price));
-        return webShopService.getAllProducts();
+    public ResponseEntity<String> addProduct(@Valid @RequestBody Product product, BindingResult br) {
+        if (br.hasErrors()) {
+            return ResponseEntity.badRequest().body("Något har blivit fel!");
+        } else {
+            webShopService.add(product);
+            return ResponseEntity.accepted().body("Det gick bra!" + product);
+        }
+
     }
 
     @PutMapping("/rest/updateprice")
-    public List<Product> updateProductPriceById(@RequestParam Long id, @RequestParam double price){
-        webShopService.updateProductPriceById(price, id);
-        return webShopService.getAllProducts();
+    public ResponseEntity<String> updateProductPriceById(@RequestParam Long id, @RequestParam double price) {
+        if (price < 1 || price > 1000) {
+            return ResponseEntity.badRequest().body("FEL! priset ska ligga mellan 1 och 1000");
+        } else {
+            webShopService.updateProductPriceById(price, id);
+            return ResponseEntity.accepted().body("Det gick utmärkt");
+        }
     }
 }
